avances requerimientos hasta el 5

+++ src/main/java/uniandes/isis2304/EPSAndes/persistencia/SQLUsuario.java
@@ -40,7 +40,7 @@ public class SQLUsuario {
 	 * @param nombre       - El nombre del tipo de bebida
 	 * @return EL número de tuplas insertadas
 	 */
-	public long adicionarUsuario(PersistenceManager pm, long idRol, String email, String nombre, int numeroDocumento, String rol, String tipoDocumento ) {
+	public long adicionarUsuario(PersistenceManager pm, long idRol, String email, String nombre, long numeroDocumento, String rol, String tipoDocumento ) {
 		Query q = pm.newQuery(SQL, "INSERT INTO " + pp.darTablaUsuario() + "(id, email, nombre, numeroDocumento, rol, tipoDocumento) values (?, ?, ?, ?, ?, ?)");
 		q.setParameters(idRol, email, nombre, numeroDocumento, rol, tipoDocumento);
 		return (long) q.executeUnique();
